// FE/src/app/core/models/enums/game/game-action-type.enum.ts
export enum GameActionType {
  CreateGame = 'CREATE_GAME',
  JoinGame = 'JOIN_GAME',
  StartGame = 'START_GAME',
  AckMulliganNotice = 'ACK_MULLIGAN_NOTICE',
  ChooseInitialPokemon = 'CHOOSE_INITIAL_POKEMON',
  PauseGame = 'PAUSE_GAME',
  ResumeGame = 'RESUME_GAME',
  DrawCard = 'DRAW_CARD',
  PlayBasicPokemon = 'PLAY_BASIC_POKEMON',
  EvolvePokemon = 'EVOLVE_POKEMON',
  AttachEnergy = 'ATTACH_ENERGY',
  PlayTrainer = 'PLAY_TRAINER',
  Retreat = 'RETREAT',
  DeclareAttack = 'DECLARE_ATTACK',
  SelectTarget = 'SELECT_TARGET',
  EndTurn = 'END_TURN',
  PromoteBenchPokemon = 'PROMOTE_BENCH_POKEMON',
  ResolveAttackChoice = 'RESOLVE_ATTACK_CHOICE',
  TakePrizeCard = 'TAKE_PRIZE_CARD',
  UseAbility = 'USE_ABILITY',
  Concede = 'CONCEDE'
}